move up - stage 1
